refactor: split SurfViewDSL.kt into logical files and clean up RenderContextDSL.kt
