+++ surf-api-bukkit/surf-api-bukkit-api/src/main/kotlin/dev/slne/surf/surfapi/bukkit/api/inventory/framework/dsl/RenderContextDSL.kt
@@ -41,7 +41,3 @@ inline fun RenderContext.resultSlot(block: @InventoryFramworkDSL BukkitItemCompo
     resultSlot().apply(block)
 }
 
-
-
-
-
