+++ surf-api-bukkit/surf-api-bukkit-api/src/main/kotlin/dev/slne/surf/surfapi/bukkit/api/inventory/framework/view/InventoryFrameworkDsl.kt
@@ -1,0 +1,5 @@
+package dev.slne.surf.surfapi.bukkit.api.inventory.framework.view
+
+@DslMarker
+@Target(AnnotationTarget.CLASS, AnnotationTarget.TYPE)
+annotation class InventoryFramworkDSL
